requirements.txt
alembic==1.18.4
uuid==1.30
sqlalchemy==2.0.49
fastapi==0.136.0
pydantic==2.13.3
pydantic_settings==2.14.0
asyncpg==0.31.0
passlib[argon2]==1.7.4
python-multipart==0.0.26
uvicorn==0.45.0
python-jose[cryptography]==3.5.0
pydantic[email]==2.13.3